move out to download.py
